moving pagedFetchResponse to domain layer
