+++ dukan_domain/src/commonMain/kotlin/net/thechance/mena/dukan/domain/repository/ProductRepository.kt
@@ -1,11 +1,16 @@
 package net.thechance.mena.dukan.domain.repository
 
 import net.thechance.mena.dukan.domain.entity.Product
+import net.thechance.mena.dukan.domain.util.PagedFetchResponse
 
 interface ProductRepository {
 
     suspend fun createProduct(product: Product)
-    suspend fun getProductsByShelfId(shelfId: String): List<Product>
+    suspend fun getProductsByShelfId(
+        shelfId: String,
+        page:Int,
+        size:Int
+    ): PagedFetchResponse<Product>
     suspend fun uploadProductImages(
         fileName: List<String>,
         fileBytes: List<ByteArray>,
